Rename MeetupsSection
Add newline to EoF in svg
